libraryか

+++ sbv2_core/src/nlp.rs
@@ -13,12 +13,12 @@ static SYMBOL_TO_ID: Lazy<HashMap<String, i32>> = Lazy::new(|| {
 pub fn cleaned_text_to_sequence(
     cleaned_phones: Vec<String>,
     tones: Vec<i32>,
-) -> (Vec<i32>, Vec<i32>, Vec<i32>) {
-    let phones: Vec<i32> = cleaned_phones
+) -> (Vec<i64>, Vec<i64>, Vec<i64>) {
+    let phones: Vec<i64> = cleaned_phones
         .iter()
-        .map(|phone| *SYMBOL_TO_ID.get(phone).unwrap())
+        .map(|phone| *SYMBOL_TO_ID.get(phone).unwrap() as i64)
         .collect();
-    let tones: Vec<i32> = tones.iter().map(|tone| *tone + 6).collect();
-    let lang_ids: Vec<i32> = vec![1; phones.len()];
+    let tones: Vec<i64> = tones.iter().map(|tone| (*tone + 6) as i64).collect();
+    let lang_ids: Vec<i64> = vec![1; phones.len()];
     (phones, tones, lang_ids)
 }
